// src/lib/pe/types.ts
export interface SectionHeader {
  name: string;
  virtualSize: number;
  virtualAddress: number;
  sizeOfRawData: number;
  pointerToRawData: number;
  characteristics: number;
}

export interface DataDirectory {
  virtualAddress: number;
  size: number;
}

export interface COFFHeader {
  machine: number;
  numberOfSections: number;
  timeDateStamp: number;
  sizeOfOptionalHeader: number;
  characteristics: number;
}

export interface ResourceLanguage {
  languageId: number;
  dataRva: number;
  dataSize: number;
  codePage: number;
}

export interface ResourceEntry {
  id: number | null;
  name: string | null;
  languages: ResourceLanguage[];
}

export interface ResourceType {
  typeId: number | null;
  typeName: string | null;
  typeLabel: string;
  entries: ResourceEntry[];
}

export interface MZHeader {
  e_cblp: number;
  e_cp: number;
  e_crlc: number;
  e_cparhdr: number;
  e_minalloc: number;
  e_maxalloc: number;
  e_ss: number;
  e_sp: number;
  e_ip: number;
  e_cs: number;
  e_lfarlc: number;
  e_ovno: number;
}

export interface PEInfo {
  dosHeader: { e_magic: number; e_lfanew: number };
  coffHeader: COFFHeader;
  optionalHeader: { magic: number; isPE32Plus: boolean; subsystem: number; dataDirectories: DataDirectory[] };
  sections: SectionHeader[];
  resources: ResourceType[] | null;
  isNE?: boolean;
  neEncoding?: string;
  isMZ?: boolean;
  mzHeader?: MZHeader;
}

export interface BitmapResult {
  id: number | null;
  name: string | null;
  languageId: number;
  bmpBlob: Blob;
  width: number;
  height: number;
  bitCount: number;
  magentaIndex: number;
  dibData: Uint8Array | null;
}
// ...
export interface BmpFileResult {
  blob: Blob;
  width: number;
  height: number;
  bitCount: number;
  magentaIndex: number;
}